feat(node): parallel IBD, storage, RPC, regtest mining; doc link fixes
- parallel_ibd: download, validation, memory, headers, feeder, prefetch, chunk assigner, checkpoints
- storage: ibd_utxo_store flush, ibd_autorepair, blockstore, chainstate, disk_utxo, serialization_cache, utxostore
- mining, rpc, mempool, network, block_processor, run_loop, sync, module api
- SECURITY.md: drop stale link after removing RPC_SECURITY_REMEDIATION_PLAN.md
- docs: fix HIGH_AVAILABILITY, RPC_REFERENCE, ZMQ_NOTIFICATIONS internal and blvm-docs src/ links
- integration tests and mainnet-ibd-config updates

+++ tests/package_relay_tests.rs
@@ -18,7 +18,7 @@ fn create_minimal_tx() -> Transaction {
     }
 }
 
-fn create_tx_with_input(prevout_hash: [u8; 32], prevout_index: u64) -> Transaction {
+fn create_tx_with_input(prevout_hash: [u8; 32], prevout_index: u32) -> Transaction {
     Transaction {
         version: 1,
         inputs: blvm_protocol::tx_inputs![TransactionInput {
